\name{gatkreport_test_v1.table}
\docType{data}
\alias{gatkreport_test_v1.table}
\title{Test table for version 1.x GATKReport}
\description{
   This is a new-style GATKReport.
}
\usage{gatkreport_test_v1.table}
\format{Text document containing multiple tables}
\source{GSA test data, Broad Institute}
\references{
   http://www.broadinstitute.org/gatk/guide/article?id=1244
}
