\name{genotype_concordance_test.table}
\docType{data}
\alias{genotype_concordance_test.table}
\title{Test table for Genotype Concordance Table Reshape}
\description{
   This is a GATKReport output by GenotypeConcordance.
}
\usage{genotype_concordance_test.table}
\format{Text document containing multiple tables}
\source{GSA test data, Broad Institute}
\references{
   http://www.broadinstitute.org/gatk/guide/article?id=1244
}
